+++ src/main/java/com/kalsym/order/service/model/OrderGroup.java
@@ -84,5 +84,8 @@ public class OrderGroup implements Serializable {
     
     @Transient
     private String shipmentName;
+    
+    @Transient
+    private String regionCountryId;
    
 }